MYR-409 Gate the Arriving rung on relevance as well as freshness
navFresh now dates the reading rather than the row, which closes the
frozen-value arm: a minutesToArrival left at 1 by a route that ended without
a cancel can no longer ride a row stamp kept current by motion.

That is not the arm this issue is named for, and a per-reading stamp alone
does not close it. A car navigating its owner's errand at the instant an
instant ride is accepted is streaming perfectly fresh navigation — one
second old, describing a route to the shops — so freshness passes, and if
the owner is two minutes from the shops the rung burns anyway. Recency and
relevance are different questions and only one of them is a timestamp on the
car.

navReadingIsOurs answers the other: a reading taken before the server told
the car where the rider is cannot be describing the rider, whatever its age.
A nil dispatch instant is not relevant either — a failed or skipped dispatch
means the car was never sent to this rider, which the dormancy gate does not
cover.

The residual is bounded and documented: dispatch is recorded when the
destination is pushed, not when the car acts on it, so a frame in that
few-second window postdates the dispatch while describing the old route.
Closing that needs the pickup coordinates, which are app-encrypted. It errs
safely where it matters — on the accept push nothing postdates the dispatch,
so the rung is left unspent for the real Arriving.

RideContext and Activity move to their own file; activity_notifier.go was
already past the 300-line cap before this change and is now well under it.
