+++ internal/push/activity_notifier.go
@@ -26,54 +26,6 @@ import (
 // columns without changing the ride's status. EndForReservationExpiry is the
 // seam for it.
 
-// Activity is one running Live Activity, the consumer-site view of a
-// go_live_activities row.
-type Activity struct {
-	RideRequestID string
-	UserID        string
-	// Token is the ActivityKit update token. P1 — never log in full.
-	Token   string
-	Sandbox bool
-	// Progress is this Activity's leg-progress anchor as last DELIVERED to the
-	// phone (MYR-398). Per-Activity rather than per-ride because the
-	// monotonicity rule it enforces is a promise made to one client about the
-	// sequence of values that client has seen.
-	Progress ProgressAnchor
-	// AlertedPhase is the highest phase this Activity's island has already been
-	// expanded for (MYR-398, migration 0029). Per-Activity for the same reason
-	// Progress is: it records what one client has been shown, and the "at most
-	// once" promise is made to that client. See activity_alert.go.
-	AlertedPhase AlertPhase
-}
-
-// RideContext is everything a content-state needs that is not the token.
-type RideContext struct {
-	// Status is the ride's lifecycle status.
-	Status string
-	// VehicleName is the car's nickname, "" when it has none.
-	VehicleName string
-	// Destination is the dropoff's short label. P1.
-	Destination string
-	// ETAMinutes is the car's carried navigation ETA in whole minutes, nil when
-	// the car has no active route.
-	ETAMinutes *int
-	// TripMilesRemaining is the car's carried remaining distance on that route,
-	// in miles (Tesla's tripDistanceRemaining), nil when it reports none. The
-	// preferred progress input — see activity_progress.go.
-	TripMilesRemaining *float64
-	// NavUpdatedAt is when the car's ROW was last written — an upper bound on
-	// the age of the two readings above, not a stamp on them. Nil for a car we
-	// have never heard from. Used ONLY to gate progress, and only as the cheap
-	// half of that gate (see navFresh / readingStalled); `eta` keeps its
-	// existing ungated behaviour.
-	NavUpdatedAt *time.Time
-	// DispatchUnderway is MYR-376's reservation-dormancy predicate as evaluated
-	// by the store: false while a reservation sleeps between accept and the
-	// earlier of its dispatch and its due instant. It gates the PICKUP leg's
-	// track — see legUnderway for what happens without it.
-	DispatchUnderway bool
-}
-
 // VehicleRide is one ride caught by an owner teardown: which ride, and the
 // status that decides how its card must be ended.
 //
